pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = [ "GenAI <csdhong@gmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-lib = "^2.1.0"
  constructs = "^10.0.5"
  python = "^3.8"

[tool.poetry.group.dev.dependencies]
projen = "0.91.1"
pytest = "7.4.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
